% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_per_spot_day.R
\name{avg_per_spot_day}
\alias{avg_per_spot_day}
\title{Break down each spot on one row per day}
\usage{
avg_per_spot_day(
  dir_csvs = "~/data/surfdata/per_spot/",
  dir_daily_spot_clean = "~/data/surfdata/per_spot_clean/"
)
}
\arguments{
\item{dir_daily_spot_clean}{}
}
\description{
Break down each spot on one row per day
}
