% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_get_data_wave.R
\name{forecast_get_data_wave}
\alias{forecast_get_data_wave}
\title{Format the wave forecast data}
\usage{
forecast_get_data_wave(d)
}
\arguments{
\item{d}{}
}
\description{
Format the wave forecast data
}
