% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_data.R
\name{make_clean_data}
\alias{make_clean_data}
\title{Make clean data for github}
\usage{
make_clean_data(dirs)
}
\arguments{
\item{dirs}{all the directories}
}
\description{
Make clean data for github
}
