% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_monthly_data.R
\name{prepare_monthly_data}
\alias{prepare_monthly_data}
\title{Prepare Monthly data}
\usage{
prepare_monthly_data(dir_daily_data = NULL, dir_monthly_data = NULL)
}
\arguments{
\item{dir_monthly_data}{}
}
\description{
Prepare Monthly data
}
