% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_per_timestamp_per_spot.R
\name{prepare_per_timestamp_per_spot}
\alias{prepare_per_timestamp_per_spot}
\title{Format all the data}
\usage{
prepare_per_timestamp_per_spot(dir_data_raw, dir_data_timestamp)
}
\description{
Format all the data
}
