% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_daily_data.R
\name{prepare_daily_data}
\alias{prepare_daily_data}
\title{Create one file for the ~8k spots per day}
\usage{
prepare_daily_data(dir_raw_download = NULL, dir_daily_data = NULL)
}
\arguments{
\item{dir}{The directory where the data is saved}
}
\description{
includes Wave height, wave rating...
}
